(PA-5715) Install xz on AIX and Solaris
This is needed so that we can decompress libxml2 and libxslt release tarballs.

On Solaris 10/11, install xz 5.2.8 to match what was done in opencsw[1].
On AIX 7.1, we install 5.2.5, because that's what we have available.
On AIX 7.2, we install 5.4.3, since other packages we're installing depend on
that version.

Also update PATH on AIX and Solaris to refer to the version we installed.

[1] https://buildfarm.opencsw.org/source/xref/opencsw/csw/mgar/pkg/xz/trunk/Makefile?r=27065