(maint) Use packaged libxml to build nokogiri on MacOS
This PR changes the gem install command for the Nokogiri gem.
Currently when building nokogiri it uses the system libxml2 and libxslt.
We now will pass in an option in the `gem install` command to use the expected
libraries.
Also on MacOS we now set the 'pkgconfig' path as an env variable when building
Nokogiri.