(PA-6422) Only install ffi gem once
Previously, we were installing the ffi gem twice, but with different arguments,
once in the base-rubygem component and again in the rubygem-ffi component. For
example:

    /opt/puppetlabs/puppet/bin/gem install --no-document --local ffi-1.15.5.gem && \
    /opt/puppetlabs/puppet/bin/gem install --no-document --local ffi-1.15.5.gem -- --enable-system-ffi
    Building native extensions. This could take a while...
    Successfully installed ffi-1.15.5
    1 gem installed
    Building native extensions with: '--enable-system-ffi'
    This could take a while...
    Successfully installed ffi-1.15.5
    1 gem installed

Now we only execute the gem install command once. On Windows, we install ffi
gems with precompiled binaries (so we don't need to specify enable/disable
flags). On non-Windows, we continue to pass enable/disable based on the ruby
version.