(PA-6422) Update rubygems patch for ruby 3.2.4
Previously, our rubygems patch wasn't being applied when cross compiling mac 12
ARM, because homebrew updated to ruby 3.2.4 and builder.rb was modified:

    - cmd = Gem.ruby.shellsplit
    + cmd = Shellwords.split(Gem.ruby)

The code for patching is very fragile and has broken multiple times without any
indication that the patch failed to apply, causing the resulting native
extensions to be compiled for the wrong architecture.

The code was also conflating the `settings[:ruby_version]` in our ruby component
with the host's ruby version (the one we're actually trying to patch). Also the
version that matters is the rubygems version (Gem::VERSION), not RUBY_VERSION.
For example, if `gem update --system` was executed during provisioning, then
our patch would fail to apply.

This commit creates a script that runs on the remote system and so can
introspect the host's ruby and rubygems versions. This way if brew updates to a
newer ruby/rubygems, we can fail the build.