(PA-6422) Namespace gem_install_options
Settings are shared across all components, so `settings[:gem_install_options]`
in the ffi gem caused its options to be passed to gems that follow like gssapi
in the bolt-runtime:

    /opt/puppetlabs/bolt/bin/gem install --no-document --local --bindir=/opt/puppetlabs/bolt/bin gssapi-1.3.1.gem -- --disable-system-libffi

This commit namespaces the options so they only apply when installing that
specific gem. The namespace is based on the full gem name like `rubygems-ffi`,
not the short name `ffi`.