Bump FFI gem from 1.15.5 to 1.16.3
Previously, we had pinned the FFI gem to 1.15.5 because 1.16.0 had a bug
that affected Windows. That bug was resolved in 1.16.3[1] and we updated
Facter (puppetlabs/facter@9ef20e5) and Puppet
(puppetlabs/puppet@aa51340) to be compatible with this version of FFI.

This commit updates all instances of FFI 1.15.5 (generic gem and
Windows 32- and 64-bit gems) to 1.16.3.

[1] https://github.com/ffi/ffi/issues/1056