(PE-38998) Downgrade rubyntlm for now
The rubyntlm gem brought in a dep on a new default gem for ruby3 (base64). This is causing issues with bolt being packaged with ruby 2.7 and jruby on 2023.x PE. The changelog does not mention any critical patches so we can take our time getting this right instead of rushing this in to the release. This commit downgrades rubyntlm to the last version before carrying the base64 dep