Bump minitar 1.0.1 for puppet 8
Minitar 1.x is not backwards compatible with 0.9, so only make this change in
agent-runtime-main for now.

Projects that depend on minitar and puppet will need to be made compatible with
minitar 1.x before bumping their puppet dependency in puppet-runtime. And when
they do bump puppet, they should bump minitar too. I added a reminder for the
latter part. This applies to:

    bolt-runtime
    pe-bolt-server-runtime-main
    pe-installer-runtime-main

pdk-runtime depends on minitar, but not puppet. When a new version of puppet
is released that depends on minitar 1.x, then modules created using an older
version of the pdk, will pull in the latest puppet, but have an incompatible
minitar version.