(PA-6881) Configure a default gem_uninstall command
With the addition logic in https://github.com/puppetlabs/puppet-runtime/pull/901 for deduplicating rexml gems an inadvertant requirement was imposed on projects to define a `gem_uninstall` command. This command should be ubiquitous (especially with impending ruby 3 only streams)  so instead of requiring all projects to configure it a default is added. This setting is still configurable at a project level, but is not required.